[286] updates tests in api package and extracts emissions factor function
